moved tests for issue 1731 into test_generators/
