+++ tests/test_generators/input/linkml_issue_1731.yaml
